fix: implement force mounting for cmdk menu items
We want some command menu items to be force-mounted (that is, they
should be mounted regardless of what the current search value is.
Examples include all search results from docs, and docs search and
AI responses when within the docs site.

Force mounting isn't implemented in a published release of `cmdk`
(https://github.com/pacocoursey/cmdk/issues/164), but we can hack it in
the interim by manipulating the item value under the hood.